+++ src/types/interfaces.ts
@@ -97,16 +97,6 @@ export interface BaseItems<T> {
     pageSize: number;
 }
 
-export interface ProductionUnits {
-    id: number;
-    name: string;
-    description: string;
-    productionUnit: ProductionUnit;
-    products: ProducerProduct[];
-    address: Address;
-    images: Image[];
-}
-
 export interface Transport {
     id: number;
     licensePlate: string;
@@ -150,6 +140,21 @@ export interface OrderItem {
     producerProduct: ProducerProduct;
     quantity: number;
     status: 'Paid' | 'Processing' | 'Shipped' | 'Delivered' | 'Canceled';
+}
+
+export interface CreateProducerProduct {
+    currentPrice: number;
+    productionDate: Date;
+    stock: number;
+    productionUnitId: number;
+    productSpecId: number;
+}
+
+export interface UpdateProducerProduct {
+    currentPrice: number;
+    productionDate: Date;
+    stock: number;
+    productionUnitId: number;
 }
 
 export type AuthenticatedUser = Consumer | Producer;